requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
tenacity
slowapi
pydantic-settings
